requirements.txt
# Knowledge Base Bot - Dependencies
# Updated for cost-effective OpenAI models (April 2025)

# Core AI/ML Dependencies
openai>=1.35.5
faiss-cpu>=1.8.0
numpy>=1.24.3

# Document Processing
pypdf>=6.0.0
pandas>=2.2.2

# Environment and Configuration
python-dotenv>=1.0.1